+++ olive/common/constants.py
@@ -20,3 +20,7 @@ DEFAULT_CACHE_DIR = ".olive-cache"
 ############# Packaging #############
 
 BASE_IMAGE = "mcr.microsoft.com/azureml/openmpi4.1.0-cuda11.8-cudnn8-ubuntu22.04"
+
+############# HF #############
+
+DEFAULT_HF_TASK = "text-generation-with-past"
