+++ olive/common/utils.py
@@ -295,27 +295,6 @@ def find_submodules(module, submodule_types, full_name=False):
     return list(submodules) if submodules else None
 
 
-def huggingface_login(token: str):
-    from huggingface_hub import login
-
-    login(token=token)
-
-
-def aml_runner_hf_login():
-    hf_login = os.environ.get("HF_LOGIN")
-    if hf_login:
-        from azure.identity import DefaultAzureCredential
-        from azure.keyvault.secrets import SecretClient
-
-        keyvault_name = os.environ.get("KEYVAULT_NAME")
-        logger.debug("Getting token from keyvault %s", keyvault_name)
-
-        credential = DefaultAzureCredential()
-        secret_client = SecretClient(vault_url=f"https://{keyvault_name}.vault.azure.net/", credential=credential)
-        token = secret_client.get_secret("hf-token").value
-        huggingface_login(token)
-
-
 def all_files(path, ignore=None):
     """Find all files in a directory recursively, optionally ignoring some paths.
