+++ olive/passes/onnx/transformer_optimization.py
@@ -9,16 +9,11 @@ from typing import TYPE_CHECKING, Any, Dict, List, Union
 
 import onnx
 
+from olive.common.hf.mappings import HIDDEN_SIZE_NAMES, MODEL_TYPE_MAPPING, NUM_HEADS_NAMES, NUM_KEY_VALUE_HEADS_NAMES
 from olive.common.utils import exclude_keys
 from olive.hardware.accelerator import AcceleratorSpec, Device
 from olive.model import ONNXModelHandler
-from olive.model.utils import (
-    HIDDEN_SIZE_NAMES,
-    MODEL_TYPE_MAPPING,
-    NUM_HEADS_NAMES,
-    NUM_KEY_VALUE_HEADS_NAMES,
-    resolve_onnx_path,
-)
+from olive.model.utils import resolve_onnx_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.pass_config import PassConfigParam
