+++ olive/passes/pytorch/sparsegpt_utils.py
@@ -12,8 +12,8 @@ import math
 import torch
 import transformers
 
+from olive.common.hf.mappings import MODELS_TO_EMBEDDINGS_MAPPING, MODELS_TO_LAYERS_MAPPING
 from olive.common.utils import get_attr
-from olive.model.utils.hf_mappings import MODELS_TO_EMBEDDINGS_MAPPING, MODELS_TO_LAYERS_MAPPING
 
 logger = logging.getLogger(__name__)
