+++ olive/passes/pytorch/torch_trt_conversion.py
@@ -9,11 +9,11 @@ from typing import Any, Dict, List, Union
 import torch
 
 from olive.common.config_utils import validate_config
+from olive.common.hf.utils import get_model_max_length
 from olive.common.utils import get_attr, tensor_data_to_device
 from olive.data.config import DataConfig
 from olive.hardware.accelerator import AcceleratorSpec, Device
 from olive.model import HfModelHandler, PyTorchModelHandler
-from olive.model.utils.hf_utils import get_model_max_length
 from olive.passes import Pass
 from olive.passes.olive_pass import PassConfigParam
 from olive.passes.pytorch.common import inherit_pytorch_from_hf
